+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Evaluate predicates and disable form submission based on `enableSubmission` form property
+
 ## [6.7.0] - 2024-09-30
 
 ### Changed